+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added `Section::$entry_count` recalculation on entry delete (Issue #16)
 - Added `Section` type attribute `visible` filter to `Sections` (Issue #15) 
 - Fixed a bug in `EntryActiveDataProvider` sorting where default orders would override the sorter
+- Enhanced `Entry::hasParentEnabled()` check
 
 ## 2.3.10 (Mar 24, 2025)